请求拦截

+++ client/src/main/java/com/ljryh/client/mapper/PermissionMapper.java
@@ -15,6 +15,6 @@ import java.util.Set;
  */
 public interface PermissionMapper extends BaseMapper<Permission> {
 
-    Set<Permission> selectPermissionByRoleId(Integer roleId);
+    Set<Permission> selectPermissionByRoleId(Long roleId);
 
 }
